requirements.txt
anyio==3.5.0
cachetools==5.0.0
certifi==2021.10.8
cfenv==0.5.3
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
deprecation==2.1.0
Flask==2.0.3
furl==2.1.3
h11==0.12.0
hdbcli==2.11.20
httpcore==0.14.7
httpx==0.22.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
orderedmultidict==1.0.1
packaging==21.3
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
rfc3986==1.5.0
sap-py-jwt==1.5.0
sap-xssec==3.3.0
six==1.16.0
sniffio==1.2.0
urllib3==1.26.8
Werkzeug==2.0.3
